+++ src/Controller/Portfolio/SkillsController.php
@@ -98,24 +98,26 @@ class SkillsController extends AbstractController
 
     #[IsGranted(new Expression('is_granted("ROLE_USER")'))]
     #[Route('/skills/{id}/delete', name: 'app_skills_delete')]
-    public function delete(Skill $skill,Connection $connection): Response
+    public function delete(Skill $skill = null , Connection $connection): Response
     {
-        try {
 
-            $sql = "DELETE FROM Skill WHERE id = :skillId";
-          
+        if (!$skill) {
+            $this->addFlash('error', 'Skill not found');
+            return $this->redirectToRoute('app_skills');
+        }
+
+        try {
+            $sql = 'DELETE FROM Skill WHERE id = :skillId';
+
             $stmt = $connection->prepare($sql);
-            $stmt->bindValue('skillId', $skill->getId() )   ;
+            $stmt->bindValue('skillId', $skill->getId())   ;
             $stmt->execute();
 
             $this->addFlash('success', 'Skill deleted successfully!');
-           
         } catch (\Exception $e) {
-            
             $this->addFlash('error', 'Error deleting skill please try again later');
-
         }
+
         return $this->redirectToRoute('app_skills');
-       
     }
 }
